Refactor tutorials and data into proper package
