+++ cmd/fyne_demo/tutorials/advanced.go
@@ -1,4 +1,4 @@
-package screens
+package tutorials
 
 import (
 	"fmt"
