+++ cmd/fyne_demo/tutorials/data.go
@@ -1,44 +1,46 @@
-package main
+package tutorials
 
 import (
 	"fyne.io/fyne"
-	"fyne.io/fyne/cmd/fyne_demo/screens"
 )
 
-type tutorial struct {
-	title, intro string
-	view         func(w fyne.Window) fyne.CanvasObject
+// Tutorial defines the data structure for a tutorial
+type Tutorial struct {
+	Title, Intro string
+	View         func(w fyne.Window) fyne.CanvasObject
 }
 
 var (
-	tutorials = map[string]tutorial{
+	// Tutorials defines the metadata for each tutorial
+	Tutorials = map[string]Tutorial{
 		"welcome": {"Welcome",
 			"Welcome to the Fyne toolkit tutorials.\nChoose an area from the options on the left.",
 			welcomeScreen,
 		},
 		"graphics": {"Graphics",
 			"See the canvas and graphics capabilities",
-			screens.GraphicsScreen,
+			GraphicsScreen,
 		},
 		"widgets": {"Widgets",
 			"Browse the toolkit widgets",
-			screens.WidgetScreen,
+			WidgetScreen,
 		},
 		"containers": {"Containers",
 			"Containers and collections",
-			screens.ContainerScreen,
+			ContainerScreen,
 		},
 		"windows": {"Windows",
 			"Window and dialog demos",
-			screens.DialogScreen,
+			DialogScreen,
 		},
 		"advanced": {"Advanced",
 			"Debug and advanced information",
-			screens.AdvancedScreen,
+			AdvancedScreen,
 		},
 	}
 
-	tutorialTree = map[string][]string{
+	// TutorialIndex  defines how our tutorials should be laid out in the index tree
+	TutorialIndex = map[string][]string{
 		"": {"welcome", "graphics", "widgets", "containers", "windows", "advanced"},
 	}
 )
