+++ cmd/fyne_demo/tutorials/graphics.go
@@ -1,4 +1,4 @@
-package screens
+package tutorials
 
 import (
 	"image/color"
