+++ cmd/fyne_demo/tutorials/theme.go
@@ -1,4 +1,4 @@
-package screens
+package tutorials
 
 import (
 	"image/color"
