+++ cmd/fyne_demo/tutorials/welcome.go
@@ -1,0 +1,58 @@
+package tutorials
+
+import (
+	"net/url"
+
+	"fyne.io/fyne"
+	"fyne.io/fyne/canvas"
+	"fyne.io/fyne/cmd/fyne_demo/data"
+	"fyne.io/fyne/container"
+	"fyne.io/fyne/layout"
+	"fyne.io/fyne/theme"
+	"fyne.io/fyne/widget"
+)
+
+func parseURL(urlStr string) *url.URL {
+	link, err := url.Parse(urlStr)
+	if err != nil {
+		fyne.LogError("Could not parse URL", err)
+	}
+
+	return link
+}
+
+func welcomeScreen(_ fyne.Window) fyne.CanvasObject {
+	a := fyne.CurrentApp()
+	logo := canvas.NewImageFromResource(data.FyneScene)
+	if fyne.CurrentDevice().IsMobile() {
+		logo.SetMinSize(fyne.NewSize(171, 125))
+	} else {
+		logo.SetMinSize(fyne.NewSize(228, 167))
+	}
+
+	return container.NewVBox(
+		layout.NewSpacer(),
+		container.NewHBox(layout.NewSpacer(), logo, layout.NewSpacer()),
+
+		container.NewHBox(layout.NewSpacer(),
+			widget.NewHyperlink("fyne.io", parseURL("https://fyne.io/")),
+			widget.NewLabel("-"),
+			widget.NewHyperlink("documentation", parseURL("https://fyne.io/develop/")),
+			widget.NewLabel("-"),
+			widget.NewHyperlink("sponsor", parseURL("https://github.com/sponsors/fyne-io")),
+			layout.NewSpacer(),
+		),
+		layout.NewSpacer(),
+
+		widget.NewCard("Choose theme", "",
+			fyne.NewContainerWithLayout(layout.NewGridLayout(2),
+				widget.NewButton("Dark", func() {
+					a.Settings().SetTheme(theme.DarkTheme())
+				}),
+				widget.NewButton("Light", func() {
+					a.Settings().SetTheme(theme.LightTheme())
+				}),
+			),
+		),
+	)
+}
