+++ cmd/fyne_demo/tutorials/window.go
@@ -1,4 +1,4 @@
-package screens
+package tutorials
 
 import (
 	"errors"
